Base URL Changes || App Name Raaise

+++ ios/ScriptTube/ViewControllers/settingControllerVC/deleteAccountVC/DeleteAccountVC.swift
@@ -22,9 +22,9 @@ class DeleteAccountVC: BaseControllerVC {
         // Do any additional setup after loading the view.
     }
     func setfonts(){
-        questionLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX30)
-        nameLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX18)
-        deactiveLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX15)
+        questionLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX22)
+        nameLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX22)
+        deactiveLbl.font = AppFont.FontName.semiBold.getFont(size: AppFont.pX14)
         messageLbl.font = AppFont.FontName.light.getFont(size: AppFont.pX12)
         submitBtn.titleLabel?.font = AppFont.FontName.medium.getFont(size: AppFont.pX17)
         profileImage.layer.cornerRadius = profileImage.frame.height / 2
@@ -39,8 +39,12 @@ class DeleteAccountVC: BaseControllerVC {
         UIApplication.keyWin!.makeKeyAndVisible()
     }
     func createDeleteAlert(){
-        let alert = UIAlertController(title: "ScripTube", message: "Are you sure you want to Delete Your Account", preferredStyle: .alert)
+        let alert = UIAlertController(title: "ScripTube", message: "Are you sure you want to Delete Your Account?", preferredStyle: .alert)
         let logOutAction = UIAlertAction(title: "Delete Account", style: .destructive){action in
+            if(!(Constant.check_Internet?.isReachable)!){
+                AlertView().showInternetErrorAlert(delegate: self)
+                return
+            }
             self.deleteAccountApi()
         }
         let cancelAction = UIAlertAction(title: "Cancel", style: .cancel)
